Fixed exporting and importing of resources
In a dev environment js files in the library are transpiled with babel so es2016 modules can be used like this:
export class ServerClient {}
import {ServerClient} from 'server/server-client'
On the other hand on the client-side thou that's not the case because the js engine does not support es2016 modules yet.
So in order the library to be used without any restrictions in nodejs environment, modules should be exported and imported
properly.
Updated README with usage example