Added namespace management in the RDF repository client.
The results are parsed via N3's DataFactory to NamedNode and finally
returned as namespace object containing the prefix and the namespace
behind it.

The RDF client uses the execute/retry mechanism of its parent class.

Added stubbed data to test the conversion to NamedNode.

Added utility class for string operations.