Implemented statement insertion in the base repo via payload (GDB-3283)
Added base class for statement payloads and another one for inserting
statements.

Added conversion methods in TermConverter for converting string subject,
predicate, object and context to appropriate terms and then to quad(s).

Added add() method in the base repository which accepts the new payload,
converts it to quads and reuses the addQuads method to send the payload
to the server as Turtle.