Reconfigured Axios to serialize query parameters in a way that can be properly read by RDF4J.
Currently Axios seralizes collection params with a [] subfix which is
not resolved by RDF4J. The serialization is now done with the QS library
which ignores such subfixes.

Added conversion for all subject, predicate, object & context method
parameters to NTRiples in case they are not already such or are not
literals (labels, numbers etc.)

Simplified repository clients to avoid resolving HTTP's response data.
It is done automatically in execute().

Updated tests to check if responses are as they should be.