Added setObjectLiteral() in add statements payload with type detection
The new method allows to set the object as literal with a data type
and language. If the data type is not provided it will autodetect it:
* integer -> xsd:integer
* floats -> xsd:decimal
* booleans -> xsd:boolean
Everything else will be considered as xsd:string

Added types.js with data type constants from the XSD schema.

Related issue https://ontotext.atlassian.net/browse/GDB-3529