Make ContentParser to accept a configuration object
Currently parsers accept a single argument but it would be better to have an object as configuration because some parsers might allow additional configuration which in current implementation is not be possible. For example N3 parsers allow a DataFactory object to be passed to their constructor.

Refactored the ContentParser to accept an object as configuration.
Updated tests.

Remove `isDefault` parameter which turns out to be redundant and update tests