Add isStreaming property to parsers
Some of the parsers are streaming by default and others don't. That's why we need a way to distinguish them in order to be able to configure the http request properly before sending it. For example if there is a configured parser for given content type and its a streaming one, then the http request must be configured with `responseType=stream` in order to have a response formatted properly and parser to be able to consume it.

Updated tests

TransactionalRepositoryClient also check if a parser isStreaming in order to configure the http request

Renamed tests