Update graphDb to 9.4.0 and fix tests
With changes in the rdf4j (https://github.com/eclipse/rdf4j/pull/2414) the ArrangedWriter class was deprecated.
As a side effect of this change now TurtleWriter-based exports have all the namespaces that are in the repository regardless of whether they are used by any of the exported statements.
This leads to the need to change the tests, although the breaking change is marked as a bug (https://github.com/eclipse/rdf4j/issues/2694).